Add files via upload
I analyse stock for 2017 and 2018 using VBA Macros in order not to recreate functions every time data is set. 2017 performed better than 2018, ENPH and RUN improved between 2017 and 2018.

As far as the execution time of the script, the refactor script runs much faster than the original script created through in the module. Refactoring the code makes it cleaner and easier to read and causes it to run faster, also refactoring the code makes it possible so i can apply the same script to 2019 and 2020 Data once i receive it.